Changed file name to bookInformation
